+++ pkg/chunk/cache/memcached_client.go
@@ -19,7 +19,7 @@ import (
 	"github.com/sony/gobreaker"
 	"github.com/thanos-io/thanos/pkg/discovery/dns"
 
-	"github.com/cortexproject/cortex/pkg/util"
+	"github.com/cortexproject/cortex/pkg/util/logutil"
 )
 
 // MemcachedClient interface exists for mocking memcacheClient.
@@ -132,7 +132,7 @@ func NewMemcachedClient(cfg MemcachedClientConfig, name string, r prometheus.Reg
 	}
 
 	if len(cfg.Addresses) > 0 {
-		util.WarnExperimentalUse("DNS-based memcached service discovery")
+		logutil.WarnExperimentalUse("DNS-based memcached service discovery")
 		newClient.addresses = strings.Split(cfg.Addresses, ",")
 	}
