+++ pkg/util/flagext/deprecated.go
@@ -3,11 +3,10 @@ package flagext
 import (
 	"flag"
 
+	"github.com/cortexproject/cortex/pkg/util/logutil"
 	"github.com/go-kit/kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
-
-	"github.com/cortexproject/cortex/pkg/util"
 )
 
 // DeprecatedFlagsUsed is the metric that counts deprecated flags set.
@@ -27,7 +26,7 @@ func (deprecatedFlag) String() string {
 }
 
 func (d deprecatedFlag) Set(string) error {
-	level.Warn(util.Logger).Log("msg", "flag disabled", "flag", d.name)
+	level.Warn(logutil.Logger).Log("msg", "flag disabled", "flag", d.name)
 	DeprecatedFlagsUsed.Inc()
 	return nil
 }
