+++ pkg/util/log.go
@@ -5,20 +5,18 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/cortexproject/cortex/pkg/util/logutil"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/weaveworks/common/logging"
-	"github.com/weaveworks/common/middleware"
 	"github.com/weaveworks/common/server"
-
-	"github.com/cortexproject/cortex/pkg/tenant"
 )
 
 var (
 	// Logger is a shared go-kit logger.
 	// TODO: Change all components to take a non-global logger via their constructors.
-	Logger = log.NewNopLogger()
+	Logger = logutil.Logger
 
 	logMessages = prometheus.NewCounterVec(prometheus.CounterOpts{
 		Name: "log_messages_total",
@@ -104,33 +102,19 @@ func (pl *PrometheusLogger) Log(kv ...interface{}) error {
 //   log := util.WithContext(ctx)
 //   log.Errorf("Could not chunk chunks: %v", err)
 func WithContext(ctx context.Context, l log.Logger) log.Logger {
-	// Weaveworks uses "orgs" and "orgID" to represent Cortex users,
-	// even though the code-base generally uses `userID` to refer to the same thing.
-	userID, err := tenant.TenantID(ctx)
-	if err == nil {
-		l = WithUserID(userID, l)
-	}
-
-	traceID, ok := middleware.ExtractTraceID(ctx)
-	if !ok {
-		return l
-	}
-
-	return WithTraceID(traceID, l)
+	return logutil.WithContext(ctx, l)
 }
 
 // WithUserID returns a Logger that has information about the current user in
 // its details.
 func WithUserID(userID string, l log.Logger) log.Logger {
-	// See note in WithContext.
-	return log.With(l, "org_id", userID)
+	return logutil.WithUserID(userID, l)
 }
 
 // WithTraceID returns a Logger that has information about the traceID in
 // its details.
 func WithTraceID(traceID string, l log.Logger) log.Logger {
-	// See note in WithContext.
-	return log.With(l, "traceID", traceID)
+	return logutil.WithTraceID(traceID, l)
 }
 
 // WithSourceIPs returns a Logger that has information about the source IPs in
