+++ pkg/util/logutil/log.go
@@ -1,0 +1,7 @@
+package logutil
+
+import "github.com/go-kit/kit/log"
+
+var (
+	Logger = log.NewNopLogger()
+)
