+++ pkg/util/spanlogger/spanlogger.go
@@ -9,7 +9,7 @@ import (
 	"github.com/opentracing/opentracing-go/ext"
 	otlog "github.com/opentracing/opentracing-go/log"
 
-	"github.com/cortexproject/cortex/pkg/util"
+	"github.com/cortexproject/cortex/pkg/util/logutil"
 )
 
 type loggerCtxMarker struct{}
@@ -26,7 +26,7 @@ type SpanLogger struct {
 
 // New makes a new SpanLogger, where logs will be sent to the global logger.
 func New(ctx context.Context, method string, kvps ...interface{}) (*SpanLogger, context.Context) {
-	return NewWithLogger(ctx, util.Logger, method, kvps...)
+	return NewWithLogger(ctx, logutil.Logger, method, kvps...)
 }
 
 // NewWithLogger makes a new SpanLogger with a custom log.Logger to send logs
@@ -35,7 +35,7 @@ func New(ctx context.Context, method string, kvps ...interface{}) (*SpanLogger, 
 func NewWithLogger(ctx context.Context, l log.Logger, method string, kvps ...interface{}) (*SpanLogger, context.Context) {
 	span, ctx := opentracing.StartSpanFromContext(ctx, method)
 	logger := &SpanLogger{
-		Logger: log.With(util.WithContext(ctx, l), "method", method),
+		Logger: log.With(logutil.WithContext(ctx, l), "method", method),
 		Span:   span,
 	}
 	if len(kvps) > 0 {
@@ -51,7 +51,7 @@ func NewWithLogger(ctx context.Context, l log.Logger, method string, kvps ...int
 // in the context. If the context doesn't have a logger, the global logger
 // is used.
 func FromContext(ctx context.Context) *SpanLogger {
-	return FromContextWithFallback(ctx, util.Logger)
+	return FromContextWithFallback(ctx, logutil.Logger)
 }
 
 // FromContextWithFallback returns a span logger using the current parent span.
@@ -68,7 +68,7 @@ func FromContextWithFallback(ctx context.Context, fallback log.Logger) *SpanLogg
 		sp = defaultNoopSpan
 	}
 	return &SpanLogger{
-		Logger: util.WithContext(ctx, logger),
+		Logger: logutil.WithContext(ctx, logger),
 		Span:   sp,
 	}
 }
